requirements.txt.txt
flask
flask_socketio
eventlet
picamera2
RPi.GPIO
pyserial
SpeechRecognition
pocketsphinx
